+++ RELEASE-NOTES.md
@@ -54,6 +54,7 @@ class Args {
 - #192 Default arity should be 1, not *, for array and collection options
 - #193 Splitting an argument should not cause max arity to be exceeded
 - #191 Arity should not limit the total number of values put in an array or collection
+- #195 Usage help should show Map types if paramLabel not specified
 - #186 Confusing usage message for collection options
 - #181 Fixed bug where incorrect help message is displayed for short options with paramLabel when arity > 1
 - #184 Improved CommandLine.setSeparator javadoc to clarify that this affects parsing only and link to the `@Command` `separator` annotation attribute.